chore(issuer): better code organization and naming
BREAKING CHANGE: Use Certificate and CertificationRequest instead of Certificate.Entity and CertificationRequest.Entity
